requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
faiss-cpu
openai
langchain
python-dotenv
undetected_chromedriver
trafilatura